Added World Interaction files;

+++ AAEmu.Game/Models/Game/World/Interactions/SummonDoodad.cs
@@ -1,5 +1,4 @@
 ﻿using AAEmu.Game.Core.Managers.UnitManagers;
-using AAEmu.Game.Models.Game.DoodadObj;
 using AAEmu.Game.Models.Game.DoodadObj.Templates;
 using AAEmu.Game.Models.Game.Skills;
 using AAEmu.Game.Models.Game.Units;
@@ -8,9 +7,10 @@ namespace AAEmu.Game.Models.Game.World.Interactions
 {
     public class SummonDoodad : IWorldInteraction
     {
-        public void Execute(Unit caster, SkillCaster casterType, BaseUnit target, SkillCastTarget targetType, uint skillId, uint doodadId, DoodadFuncTemplate objectFunc)
+        public void Execute(Unit caster, SkillCaster casterType, BaseUnit target, SkillCastTarget targetType,
+            uint skillId, uint doodadId, DoodadFuncTemplate objectFunc)
         {
-            var doodad = DoodadManager.Instance.Create(0, (uint)doodadId, caster);
+            var doodad = DoodadManager.Instance.Create(0, doodadId, caster);
             doodad.Position = target.Position.Clone();
             doodad.Spawn();
         }
